Rename bot to core
