+++ core/classes/bot-settings.js
