+++ core/classes/command.js
