+++ core/classes/guild.js
