+++ core/classes/owner.js
