+++ core/classes/permission.js
