+++ core/classes/plugin.js
