+++ core/classes/spam-user.js
