+++ core/classes/user-role.js
