+++ core/utils/database-helper.js
