+++ core/utils/file-helper.js
